% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejoints.R
\name{dejoint_1}
\alias{dejoint_1}
\title{Sample from empirical joint density}
\usage{
dejoint_1(x, input, sigma = 1)
}
\arguments{
\item{x}{the point to be evaluated}

\item{input}{data in matrix form}

\item{sigma}{the standard deviation used to construct the empirical}
}
\description{
Sample from empirical joint density
}
