% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejoints.R
\name{rejoint}
\alias{rejoint}
\title{Sample from empirical joint density}
\usage{
rejoint(n, input, sigma = 1)
}
\arguments{
\item{n}{number of samples required}

\item{input}{data in matrix form}

\item{sigma}{the standard deviation used to construct the empirical density}
}
\description{
Sample from empirical joint density
}
